FEAT[logic]: ADD basic logic for lightning mode
Made the colors lighter until white using probably an
overcomplicated method of converting from rgb to hsl so
the process of making the color lighter is easier. althouught
it ended up being more complex than what probably was needed
maybe following a simple approach of transparency.
It's the basic logic. Need to refactor